admin panel: fix build
